[WINESYNC] mshtml: Merge htmlstyle3.c into htmlstyle.c.

wine commit id f7597a6a1f5c156316b896f22aa1f6ea6a558c06 by Jacek Caban <jacek@codeweavers.com>
