+++ sdk/tools/winesync/mshtml.cfg
@@ -8,4 +8,4 @@ files:
   include/mshtmhst.idl: sdk/include/psdk/mshtmhst.idl
   include/mshtml.idl: sdk/include/psdk/mshtml.idl
 tags:
-  wine: d22303703428842106f09432111785af7f5b1ad0
+  wine: f7597a6a1f5c156316b896f22aa1f6ea6a558c06
